Add inline button to update status messages (#354)
* Add abstraction for representing Telegram messages with metadata

This also adds several helper functions to send such messages more
easily.
Using the abstraction makes sense if the "content" of a message is
defined before we know how/where it is to be sent.

* Add possibility to use inline button to manually update status messages

* [fix] Correctly update photo when updating an existing progress notification

In my refactor, it seems like I accidentally forgot to pass the photo on
to the method updating the message. Whoops.